feat(substrate,component): widen kind id u32 → u64 (ADR-0030 Phase 1) (#132)
## Summary

Phase 1 of ADR-0030: mechanical `u32 → u64` widen of every kind-id
touchpoint. No behavior change — ids are still sequentially assigned by
the registry. Phase 2 will swap the sequential assignment for
`fnv1a(name ++ postcard(schema))` and retire `resolve_kind_p32`.

Mirrors the shape of ADR-0029's split (PR #129 widened mailbox ids
before PR #130 switched to hashing).

**Touchpoints**

- `MailKind = u64`, `KIND_NOT_FOUND = u64::MAX`
- FFI: `send_mail_p32` / `reply_mail_p32` kind params widen to i64;
`resolve_kind_p32` result widens to i64
- `Registry::{kind_id, kind_name, kind_descriptor, register_kind*}`
return/accept u64
- Guest SDK: `KindId<K>.raw`, `Sink<K>.kind`, `Mail::kind`, `KindTable`
entries → u64
- `Component::receive` typed func widens to `(u64, u32, u32, u32) ->
u32`
- Every `receive_p32` WAT shim's first param widens `i32 → i64`; WAT
`send_mail` / `reply_mail` call sites use `i64.const` for the kind arg
- `aether-substrate/main.rs`'s five cached input-kind ids widen to u64

## Test plan

- [x] `cargo check --workspace --all-targets`
- [x] `cargo test --workspace` — all green
- [x] `cargo clippy --all-targets -- -D warnings`
- [x] `cargo fmt -- --check`

+++ crates/aether-component/src/lib.rs
@@ -52,7 +52,8 @@ pub use kinds::{Cons, KindList, KindTable, Nil};
 
 /// Sentinel returned by `raw::resolve_kind` when the substrate has not
 /// registered the requested kind name. Mirrors the host constant.
-pub const KIND_NOT_FOUND: u32 = u32::MAX;
+/// Widened to `u64` in ADR-0030 Phase 1.
+pub const KIND_NOT_FOUND: u64 = u64::MAX;
 
 /// Phantom-typed wrapper around a resolved kind id. A `KindId<Tick>`
 /// cannot be passed where a `KindId<DrawTriangle>` is expected — the
@@ -63,7 +64,7 @@ pub const KIND_NOT_FOUND: u32 = u32::MAX;
 /// `kind` parameters in a hand-rolled `receive` shim (ADR-0014's
 /// `Mail::decode` will make the raw-int compare go away).
 pub struct KindId<K: Kind> {
-    raw: u32,
+    raw: u64,
     _k: PhantomData<fn() -> K>,
 }
 
@@ -82,14 +83,14 @@ impl<K: Kind> Eq for KindId<K> {}
 
 impl<K: Kind> KindId<K> {
     /// The raw kind id the substrate assigned. Exposed for hand-rolled
-    /// receive shims that `match` on the inbound `kind: u32` parameter.
-    pub fn raw(self) -> u32 {
+    /// receive shims that `match` on the inbound `kind: u64` parameter.
+    pub fn raw(self) -> u64 {
         self.raw
     }
 
     /// Returns `true` if `raw` is the id the substrate assigned to `K`.
     /// Convenience over `kind_id.raw() == raw`.
-    pub fn matches(self, raw: u32) -> bool {
+    pub fn matches(self, raw: u64) -> bool {
         self.raw == raw
     }
 }
@@ -102,7 +103,7 @@ impl<K: Kind> KindId<K> {
 /// Built via `resolve_sink::<K>(name)` during init.
 pub struct Sink<K: Kind> {
     mailbox: u64,
-    kind: u32,
+    kind: u64,
     _k: PhantomData<fn() -> K>,
 }
 
@@ -121,7 +122,7 @@ impl<K: Kind> Sink<K> {
     }
 
     /// Raw kind id. Exposed for the same reason as `mailbox`.
-    pub fn kind(self) -> u32 {
+    pub fn kind(self) -> u64 {
         self.kind
     }
 }
@@ -490,7 +491,7 @@ impl<'a> PriorState<'a> {
 /// linear memory (the substrate placed them there before the FFI
 /// call), so zero-copy is possible when alignment permits.
 pub struct Mail<'a> {
-    kind: u32,
+    kind: u64,
     // Stored as `usize` so `Mail::decode` can reconstruct a full host
     // pointer for tests, while the FFI path (`__from_raw`) widens the
     // incoming `u32` address. On wasm32 `usize == u32` so this is a
@@ -511,7 +512,7 @@ impl<'a> Mail<'a> {
     /// delivers `ptr` as a wasm32 offset (`u32`); this widens it.
     #[doc(hidden)]
     pub unsafe fn __from_raw(
-        kind: u32,
+        kind: u64,
         ptr: u32,
         count: u32,
         sender: u32,
@@ -531,7 +532,7 @@ impl<'a> Mail<'a> {
     /// from a host pointer go through here so 64-bit addresses survive.
     #[doc(hidden)]
     #[cfg(test)]
-    unsafe fn __from_ptr_test(kind: u32, ptr: usize, count: u32, sender: u32) -> Self {
+    unsafe fn __from_ptr_test(kind: u64, ptr: usize, count: u32, sender: u32) -> Self {
         Mail {
             kind,
             ptr,
@@ -547,7 +548,7 @@ impl<'a> Mail<'a> {
     #[doc(hidden)]
     #[cfg(test)]
     unsafe fn __from_ptr_test_with_table(
-        kind: u32,
+        kind: u64,
         ptr: usize,
         count: u32,
         sender: u32,
@@ -566,7 +567,7 @@ impl<'a> Mail<'a> {
     /// Raw kind id the substrate routed this mail under. Match against
     /// a cached `KindId<K>` via `kind_id.matches(mail.kind())`, or use
     /// `decode::<K>(kind_id)` and let it be the discriminator.
-    pub fn kind(&self) -> u32 {
+    pub fn kind(&self) -> u64 {
         self.kind
     }
 
@@ -779,7 +780,7 @@ macro_rules! export {
         /// mail with no meaningful reply target. Exported under the
         /// `_p32` suffix per ADR-0024 Phase 1.
         #[unsafe(export_name = "receive_p32")]
-        pub unsafe extern "C" fn receive(kind: u32, ptr: u32, count: u32, sender: u32) -> u32 {
+        pub unsafe extern "C" fn receive(kind: u64, ptr: u32, count: u32, sender: u32) -> u32 {
             let Some(instance) = (unsafe { __AETHER_COMPONENT.get_mut() }) else {
                 return 1;
             };
